requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
autograd==1.8.0
autograd-gamma==0.5.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.62.0
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.116.1
Flask==3.1.1
fonttools==4.59.0
formulaic==1.2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
interface-meta==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lifelines==0.30.0
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
mlflow==3.2.0
mlflow-skinny==3.2.0
mlflow-tracing==3.2.0
narwhals==2.0.1
ngboost==0.5.6
numba==0.61.2
numpy==2.2.6
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
patsy==1.0.1
pillow==11.3.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.5.3
starlette==0.47.2
statsmodels==0.14.5
stumpy==1.13.0
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
tsfresh==0.21.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
python-multipart==0.0.20
prometheus_client==0.17.1
